DAOS-6225 control: Parse server access_points as hostlist (#4129)
Dedupe and normalize before further validation in order
to ensure that the access_points list contains a set of
distinct and valid entries.
