+++ src/control/server/config/server.go
@@ -27,6 +27,7 @@ import (
 	"context"
 	"fmt"
 	"io/ioutil"
+	"net"
 	"os"
 	"path/filepath"
 	"strconv"
@@ -438,6 +439,10 @@ func (c *Server) Validate(log logging.Logger) (err error) {
 		return FaultConfigNoProvider
 	}
 
+	c.AccessPoints, err = common.ParseHostList(c.AccessPoints, c.ControlPort)
+	if err != nil {
+		return errors.Wrap(err, "unable to parse access_points")
+	}
 	switch {
 	case len(c.AccessPoints) < 1:
 		return FaultConfigBadAccessPoints
@@ -445,9 +450,8 @@ func (c *Server) Validate(log logging.Logger) (err error) {
 		return FaultConfigEvenAccessPoints
 	}
 
-	for i := range c.AccessPoints {
-		// apply configured control port if not supplied
-		host, port, err := common.SplitPort(c.AccessPoints[i], c.ControlPort)
+	for _, ap := range c.AccessPoints {
+		host, port, err := net.SplitHostPort(ap)
 		if err != nil {
 			return errors.Wrap(FaultConfigBadAccessPoints, err.Error())
 		}
@@ -460,8 +464,6 @@ func (c *Server) Validate(log logging.Logger) (err error) {
 		if port == "0" {
 			return FaultConfigBadControlPort
 		}
-
-		c.AccessPoints[i] = fmt.Sprintf("%s:%s", host, port)
 	}
 
 	netCtx, err := netdetect.Init(context.Background())
